handling tx buffer exceptions

+++ src/provider_can/can/can_dispatcher.h
@@ -51,7 +51,7 @@ typedef struct {
   // The number attributed to the method to be accessed
   uint8_t method_number;
   // the value to pass as a parameter to the method
-  float parameter_value;
+  double parameter_value;
   // String to pass for additionnal parameters
   std::string string_param;
 } ComputerMessage;
